requirements.txt
ansicon==1.89.0
asgiref==3.7.2
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
codefind==0.1.3
Django==4.2.3
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.6
GDAL @ file:///C:/Users/user/pipwin/GDAL-3.4.3-cp39-cp39-win_amd64.whl
gis==0.2.1
idna==3.4
inflection==0.5.1
jinxed==1.2.0
Js2Py==0.74
jurigged==0.5.5
ovld==0.3.2
packaging==23.1
Pillow==10.0.0
pipwin==0.5.2
psycopg2==2.9.6
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.3
watchdog==3.0.0
wcwidth==0.2.6
